+++ 0x03-debugging/main.h
@@ -3,7 +3,7 @@
 
 #include <stdio.h>
 
-int largest_number(int a, int b, int c);
+int largest_number(long a, long b, long c);
 void positive_or_negative(int i);
 
 #endif /* MAIN_H */
